Create Stage 3 - 2 Nema17 Motor Controlled By Input
This code controls two NEMA17 motors by user input. Type a number 0-190 in the serial monitor and it will output the speed at which it is going.
When testing, the NEMA17 motors were unable to change directions when calling "forward()" and "backward()" I don't know why, I'm assuming it's because there was not enough power to power the motors properly.